% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{getProxy}
\alias{getProxy}
\title{Sample a random proxy address}
\usage{
getProxy(proxies, filter = c())
}
\arguments{
\item{proxies}{data frame containing proxy addresses and ports. Returned by 'refreshProxies()' function}

\item{filter}{vector containing addresses of proxies that were already tried but failed to resolve.}
}
\value{
named vector where: the value is the port and the name is the proxy address
}
\description{
This function takes the list of proxies and randomly samples one of the addresses. The list can be filtered for proxies that were already tried but failed.
}
