% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{loadDatabase}
\alias{loadDatabase}
\title{Load database if exists}
\usage{
loadDatabase()
}
\value{
If the database 'listings.rds' exists, the function returns this database. Else, it will return an empty tibble containing the same columns.
}
\description{
Load database if exists
}
