% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\docType{class}
\name{Pararius-class}
\alias{Pararius-class}
\alias{.Pararius}
\alias{Pararius}
\title{Pararius class}
\usage{
Pararius(city, min_price, max_price, blacklist = list())
}
\arguments{
\item{city}{city you want to look for listings}

\item{min_price}{minimum price for the listing}

\item{max_price}{maximum price for the listing}

\item{blacklist}{list of realtors to blacklist}
}
\description{
This class is used to scrape the Pararius website

Initiate an object of class Pararius
}
\section{Slots}{

\describe{
\item{\code{city}}{city you want to look for listings}

\item{\code{min_price}}{minimum price for the listing}

\item{\code{max_price}}{maximum price for the listing}

\item{\code{blacklist}}{list of realtors to blacklist}
}}

