% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{refreshProxies}
\alias{refreshProxies}
\title{Download and save a list of proxies}
\usage{
refreshProxies()
}
\value{
TRUE if new proxy list has been downloaded. FALSE if not.
}
\description{
This function downloads a fresh list of https proxies every 24 hours. This is necessary to load the Pararius/Funda websites because they have good anti-scraping protection.
}
