% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{requestPage}
\alias{requestPage}
\title{Request page}
\usage{
requestPage()
}
\value{
httr request information if successful. Raises error if unsuccessful.
}
\description{
This function loads the Pararius/Funda page.
}
