% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\docType{methods}
\name{update_rentals}
\alias{update_rentals}
\alias{update_rentals,Pararius-method}
\title{Download, parse and process new listings from an aggregator}
\usage{
update_rentals(x)

\S4method{update_rentals}{Pararius}(x)
}
\arguments{
\item{x}{a Pararius or Funda object}
}
\value{
a data frame containing new listings. Also saves RDS databases for proxies & listings in the current working directory.
}
\description{
Download, parse and process new listings from an aggregator
}
